requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
sympy
fastapi-utils
python-jose[cryptography]
passlib
PyJWT
sympy
requests
apscheduler
motor
httpx